ai-review/requirements.txt
fastapi
uvicorn
tensorflow
numpy
python-mecab-ko
scikit-learn
